Enumeration GlowFieldFlags had member GlowFieldFlag_All defined to be 0xffffffff. For some variants of gcc, this will be expanded to 64bit in order to represent this 32bit number, since enumeration defaults to signed integer.
This becomes a problem when incomming signed 32bit integer flags, are typecasted, causing sign extension (value will stored is then 0xffffffffffffffff), so parsers that checks if packet.fields == GlowFieldFlag_All fails, since 0xffffffffff != 0xffffffffffffffff.

Easiest fix for now is this patch for now is this patch.

Best permanent fix will probably to define GlowFieldFlags as a uint32_t, and the different flags as #define, to ensure bitsize and avoid sign-extensions. Please enter the commit message for your changes. Lines starting